Replace `reinstall` script with a simpler `postinstall`
These changes are made for a couple of reasons:

**Ideally running `npm install`**

`npm install` is the de-facto way of installing dependencies in npm projects.

The built-in install script has lifecycle scripts that allows us to perform
actions *after* all npm dependencies has been installed for the project.

npm ensures that dependencies installed which includes executable files
are available in `$PATH` when npm scripts are run. In practise that means
a custom `build` npm script might use `elm` as if it were globally installed
on the machine, but it infact is a `devDependency` to that particular project.

**Avoiding global dependencies**

Ideally a project does not require global packages installed on the machine
to run. This is cause of two reasons basically:

- There is no good way to specify a system wide `dependency` and its version
- A developer might want to run different versions of e.g. `elm` on different
  projects. By enforcing global installs, that's quite painful.